refactor naming file for component
