+++ src/containers/ContactList.container.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import filterFunction from '../../lib/filterFunction';
-import ContactList from '../components/ContactList';
+import ContactList from '../components/ContactList.component';
 
 const mapStateToProps = state => ({
   data: filterFunction(state.contacts, state.filterKeyword)
